requirements.txt
black==22.8.0
click==8.1.3
colorama==0.4.5
mypy-extensions==0.4.3
numpy==1.23.3
pandas==1.4.4
pathspec==0.10.1
platformdirs==2.5.2
psutil==5.9.2
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
tomli==2.0.1
